Add `spack path` command, defaulting to install directory
`spack location` is verbose, and its default (the `spack-source` dir)
is almost never what users want. If I pull `spack location` usages
on public GitHub usage, this is what I get:

    arg                  count     pct
    ------------------------------------
    -i                     325   61.4%
    --install-dir          179   33.8%
    (spec, no flag)          7    1.3%
    (no args)                3    0.6%
    --first                  3    0.6%
    --repo / -r / -e / -p   <=2  <0.5% each

So, from 529 occurrences across 306 files in 203 repos,
`spack location -i` is 95.3% of calls.

This PR adds `spack path`, which is for all intents and purposes the
same as `spack location`, but defaults to the install directory (`-i`)
when no directory flag is given, so:

    spack path <spec> == spack location -i <spec>

`spack location` is deprecated, but not in a noisy way. If you run
`spack location -h`, you get:

```
Warning: spack location is deprecated; use spack path instead.

spack path takes all the same options as spack location, but it prints the
install prefix (-i) by default instead of the source directory (-c).
```

Since there are many scripts written using `spack location`, we should
keep it around at least until a potential Spack 2.0. But we can update the
tutorial and docs to use `spack path`.

This PR does *not* change the behavior of `spack cd`. Querying github data
for `spack cd`, about 50% of calls are arg-less and 50% take arguments.
`spack cd` *seems* to mostly be used in developer workflows, where the
default of switching to the source directory makes sense. `spack location`
is used scripts that need to refer to spack installation locations/paths.

- [x] Add `spack path` command
- [x] `spack.cmd.location` now delegates to `spack.path`
- [x] update tests for new change
