+++ lib/spack/spack/cmd/cd.py
@@ -5,7 +5,7 @@
 import argparse
 
 import spack.cmd.common
-import spack.cmd.location
+import spack.cmd.path
 
 description = "cd to spack directories in the shell"
 section = "user environment"
@@ -16,11 +16,11 @@ def setup_parser(subparser: argparse.ArgumentParser) -> None:
     """This is for decoration -- spack cd is used through spack's
     shell support.  This allows spack cd to print a descriptive
     help message when called with -h."""
-    spack.cmd.location.setup_parser(subparser)
+    spack.cmd.path.setup_parser(subparser)
 
 
 def cd(parser, args):
     spec = " ".join(args.spec) if args.spec else "SPEC"
-    spack.cmd.common.shell_init_instructions(
-        "spack cd", "cd `spack location --install-dir %s`" % spec
-    )
+    # `spack cd` resolves directories through `spack location`, which defaults
+    # to a spec's source dir (-c) when no directory option is given.
+    spack.cmd.common.shell_init_instructions("spack cd", "cd `spack location %s`" % spec)
